Fixed styling and homepage
