+++ src/App.jsx
@@ -2,7 +2,6 @@ import React,{useState, useEffect} from 'react';
 import { Routes, Route } from "react-router-dom";
 import axios from 'axios';
 import HomePage from "./components/HomePage";
-import LandingPage from "./components/LandingPage";
 import Navbar from "./components/Navbar";
 import Login from "./components/Login";
 import SignUp from "./components/SignUp";
@@ -36,9 +35,7 @@ useEffect(() => {
       */}
       <Navbar/>
       <Routes>
-        <Route path="/" element={<HomePage />} />
-      
-        <Route path ="/getstarted" element={<LandingPage />} />
+        <Route path ="/" element={<HomePage />} />
      
         <Route path ="/signup" element={<SignUp disabled ={disabled} setDisabled={setDisabled}/>} />
