+++ src/styles/App.css
@@ -10,6 +10,7 @@
     display: flex;
     justify-content: space-between;
     margin:1%;
+    animation: 3s ease-in content;
     
 }
 .menu{
@@ -24,6 +25,7 @@
 }
 .homepage-container{
     margin:10%;
+    width:100%;
     padding:0% 2%;
     display: flex;
     flex-direction: column;
@@ -49,11 +51,9 @@ hr{
     margin:2% 9%;
     display: flex;
     flex-direction: column;
-    
+    animation: 1s ease-in content;
 }
-.landing-container h1{
-    padding:10%;
-}
+
 .landingpage-content{
     display: flex;
     justify-content: space-between;  
@@ -79,6 +79,7 @@ hr{
     margin: 6rem auto 8.1rem auto;
     width: 30%;
     padding: 2% 3%;
+    animation: 0.5s ease-in content;
 }
 #signUp-form{
     width: 39%;
@@ -94,6 +95,7 @@ hr{
     margin: 6rem auto 8.1rem auto;
     width: 300px;
     padding: 12px 44px;
+    animation: 1s ease-in content;
 }
 .input-container{
     display: flex;
@@ -109,6 +111,11 @@ input{
     width: 95%;
     height:2.1rem;
 }
+input:focus, textarea:focus{
+    background-color: rgb(19, 26, 49);
+    outline: none;
+    border: 1px solid rgb(2, 189, 117);
+}
 
 /* Animations */
 @keyframes slideInLeft {
@@ -121,3 +128,11 @@ input{
      transform: translateX(0)
     }
   }
+@keyframes content{
+    0%{
+        opacity: 0;
+    }
+    100%{
+        opacity: 1;
+    }
+}
